LayoutTests/canvas/philip/tests/size.attributes.idl.set.zero-expected.txt

< [index] >
size.attributes.idl.set.zero
Setting width/height IDL attributes to 0
References: size.width size.height
Actual output:
Passed
